pyproject.toml
[tool.poetry]
name = "line_trnsltrbt"
version = "0.1"
description = ""
authors = ["Wayne Sun <x@tssrct.us>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
google-cloud-translate = "^3.15.0"
gunicorn = "^21.2.0"
psutil = "^5.9.0"
line-bot-sdk = "3.17.0"
flask = "^3.1.0"
flask-cors = "^5.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"